Introspection
Even though generally not recommended, I used introspection to easily find out what the user typed int the cmd and how to respond to it.

Except for the functions outside the class where I used the good ol' elif.

Also I literally changed everything that had 'balance' in it with 'debt' (same with Balance and Debt)

Things to do:
-make sure this works with spreadsheet that aren't
limited (lotta empty cols and rows)
-add a function to change the position of two columns based using names as inputs
-add a function that that deletes a column using
a name as an input.